fastapi_template/requirements.txt
fastapi
alembic
uvicorn
pydantic-settings
# pydantic
pydantic[email]
mysqlclient
sqlalchemy
pymysql
python-jose
python-multipart
tenacity